+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Main
 
 ##### New Features/Enhancements
+- Add Swift Package Manager Support
+  [notbenoit](https://github.com/notbenoit)
+  [#73](https://github.com/Fueled/ios-utilities/pulls/73)
 
 - Add `ActionProtocol`
 - Add `AnyAction`, allowing to type-erase any actions represented by a `ActionProtocol`